+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased](https://github.com/orisai/nette-monolog/compare/1.5.0...v1.x)
 
+### Changed
+
+- Composer
+	- Allow PHP 8.4
+
 ## [1.5.0](https://github.com/orisai/nette-monolog/compare/1.4.0...1.5.0) - 2024-07-09
 
 ### Changed
@@ -31,9 +36,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [1.3.1](https://github.com/orisai/nette-monolog/compare/1.3.0...1.3.1) - 2024-06-21
 
-### Added
+### Changed
 
-- Allow PHP 8.3
+- Composer
+	- Allow PHP 8.3
 
 ## [1.3.0](https://github.com/orisai/nette-monolog/compare/1.2.6...1.3.0) - 2023-01-13